Make ReferenceType just a base class for the actual reference types: ArrayType and ClassOrInterfaceType
