+++ javaparser-core/src/main/java/com/github/javaparser/ast/type/ReferenceType.java
@@ -21,67 +21,17 @@
  
 package com.github.javaparser.ast.type;
 
-import java.util.List;
-
 import com.github.javaparser.Range;
-import com.github.javaparser.ast.expr.AnnotationExpr;
-import com.github.javaparser.ast.nodeTypes.NodeWithAnnotations;
-import com.github.javaparser.ast.nodeTypes.NodeWithType;
-import com.github.javaparser.ast.visitor.GenericVisitor;
-import com.github.javaparser.ast.visitor.VoidVisitor;
 
 /**
  * @author Julio Vilmar Gesser
  */
-public class ReferenceType extends Type<ReferenceType> implements NodeWithType<ReferenceType>, NodeWithAnnotations<ReferenceType> {
-
-	private Type type;
+public abstract class ReferenceType<T extends ReferenceType> extends Type<T> {
 
     public ReferenceType() {
 	}
 
-	public ReferenceType(final Type type) {
-		setType(type);
-	}
-
-	public ReferenceType(final Range range, final Type type) {
+	public ReferenceType(final Range range) {
 		super(range);
-		setType(type);
-	}
-
-    public ReferenceType(Range range, Type type, List<AnnotationExpr> annotations) {
-        super(range, annotations);
-        setType(type);
-    }
-
-	/**
-	 * Creates a new {@link ReferenceType} for a class or interface.
-	 *
-	 * @param name
-	 *            name of the class or interface
-	 * @return instanceof {@link ReferenceType}
-	 */
-	public static ReferenceType create(String name) {
-		return new ReferenceType(new ClassOrInterfaceType(name));
-	}
-
-	@Override public <R, A> R accept(final GenericVisitor<R, A> v, final A arg) {
-		return v.visit(this, arg);
-	}
-
-	@Override public <A> void accept(final VoidVisitor<A> v, final A arg) {
-		v.visit(this, arg);
-	}
-
-	@Override
-	public Type getType() {
-		return type;
-	}
-
-	@Override
-    public ReferenceType setType(final Type type) {
-		this.type = type;
-		setAsParentNodeOf(this.type);
-        return this;
 	}
 }
