+++ javaparser-core/src/main/java/com/github/javaparser/ast/visitor/GenericVisitor.java
@@ -125,8 +125,6 @@ public interface GenericVisitor<R, A> {
 
 	R visit(PrimitiveType n, A arg);
 
-	R visit(ReferenceType n, A arg);
-	
 	R visit(ArrayType n, A arg);
 	
 	R visit(DimensionedArrayType n, A arg);
