+++ javaparser-core/src/main/java/com/github/javaparser/ast/visitor/VoidVisitor.java
@@ -125,8 +125,6 @@ public interface VoidVisitor<A> {
 
 	void visit(PrimitiveType n, A arg);
 
-	void visit(ReferenceType n, A arg);
-
 	void visit(ArrayType n, A arg);
 	
 	void visit(DimensionedArrayType n, A arg);
