+++ javaparser-core/src/main/java/com/github/javaparser/ast/visitor/VoidVisitorAdapter.java
@@ -592,12 +592,6 @@ public  class VoidVisitorAdapter<A> implements VoidVisitor<A> {
 		n.getQualifier().accept(this, arg);
 	}
 
-	@Override public void visit(final ReferenceType n, final A arg) {
-		visitComment(n.getComment(), arg);
-		visitAnnotations(n, arg);
-		n.getType().accept(this, arg);
-	}
-
 	@Override
 	public void visit(ArrayType n, A arg) {
 		visitComment(n.getComment(), arg);
